requirements.txt
requests==2.26.0
python-dotenv~=1.0.0
pandas~=2.0.3